executed_sql_resource, executed_sql_resource_data 파티셔닝하기위해 컬럼 조절.

+++ com.hangum.tadpole.commons/src/com/hangum/tadpole/commons/dialogs/message/dao/RequestResultDAO.java
@@ -22,7 +22,7 @@ import com.hangum.tadpole.commons.libs.core.define.PublicTadpoleDefine;
  */
 public class RequestResultDAO {
 	
-	int seq;
+	long seq;
 	
 	/** Execute start time */
 	Timestamp startDateExecute;
@@ -110,11 +110,11 @@ public class RequestResultDAO {
 		this.dbSeq = dbSeq;
 	}
 
-	public int getSeq() {
+	public long getSeq() {
 		return seq;
 	}
 
-	public void setSeq(int seq) {
+	public void setSeq(long seq) {
 		this.seq = seq;
 	}
